+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 2.15.0 - 2022-03-23
+
+### Added
+- Added new prop `paramKey` to `QasListItems`.
+
+### Changed
+- `QasResizer` changed baseURL `'https://d17ouzaofz81f3.cloudfront.net/'` to `'https://image-resize.nave.dev/'`
+
+## 2.14.1 - 2022-03-21
+
+### Fixed
+- Removed `this.file = []` in method `__addFiles` in `QasCustomUploader`.
+
+## 2.14.0 - 2021-02-11
+
+### Added
+- Added new prop `title` to `QasAppMenu`.
+
+### Changed
+- Logic of handling modules inside `QasAppMenu`.
+
+### Removed
+- Removed prop `currentModule` from `QasAppMenu`, now the "currentModule" it's handled inside the component (it's not a breaking change).
+
 ## 2.13.0 - 2021-01-12
 
 ### Added